requirements.txt
python-dotenv
flask
flask_cors
langchain
tiktoken
openai
chromadb
gradio
